requirements.txt
aiogram==2.25.2
aiohttp==3.8.6
aiosignal==1.3.1
anyio==4.6.0
async-timeout==4.0.3
attrs==24.2.0
Babel==2.9.1
certifi==2024.8.30
charset-normalizer==3.4.0
exceptiongroup==1.2.2
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
magic-filter==1.0.12
multidict==6.1.0
propcache==0.2.0
psycopg2-binary==2.9.9
python-telegram-bot==21.6
pytz==2024.2
sniffio==1.3.1
typing_extensions==4.12.2
yarl==1.14.0
